pyproject.toml
[project]
name = "proto-pipe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "duckdb>=1.5.1",
    "mypy>=1.19.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pytz>=2026.1.post1",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
]
